Generate json formatted dependencies for dep package
With this change we know output a json formatted list of dependencies for a
debian package given the url to the debian page for the package. For example, to
get the dependencies for vlc one would use the following:

`scraper https://packages.debian.org/jessie/web/vlc`

The output to stdout would like like such...

```
...
{
    "name": "fonts-freefont-ttf",
    "config-opts": "",
    "sources": [
      {
        "type": "archive",
        "url": "http://http.debian.net/debian/pool/main/v/vlc/vlc_2.2.4.orig.tar.xz",
        "sha256": "1632e91d2a0087e0ef4c3fb4c95c3c2890f7715a9d1d43ffd46329f428cf53be"
      }
    ]
  },
  {
    "name": "gcc-4.9-base",
    "config-opts": "",
    "sources": [
      {
        "type": "archive",
        "url": "http://http.debian.net/debian/pool/main/g/gcc-4.9/gcc-4.9_4.9.2.orig.tar.gz",
        "sha256": "861aa811d5f9e9ecf32d8195d2346fc434eba7e17330878ed3d876c49a32ec4e"
      }
    ]
  },
  {
    "name": "libgcc1",
    "config-opts": "",
    "sources": [
      {
        "type": "archive",
        "url": "http://http.debian.net/debian/pool/main/g/glibc/glibc_2.19.orig.tar.xz",
        "sha256": "746e52bb4fc9b2f30bcd33d415172a40ab56c5fff6c494052d31b0795593cc60"
      }
    ]
  },

...
```

The output is designed to be used in the modules section of a Flatpak json file.
It can be used as a starting point to build Flatpak packages based on tarballs
from Debian. One then needs to go through and remove any dependencies that are
already found in the Sdk that is being used.

Duplicates are filtered out. Ofter times, different Debian packages are derived
from the same "original" tarball. Those are also filtered out. Although, this
might be something to revert.

If a package does not have an "original" tarball then the fields are left empty.