Support creating dep list using .deb packages
With this change we can opt for getting the Debian package information. So,
instead of building this from the "orig.tar" files we can get the binary debs.

This also adds some new flags. The "arch" flag allows you use packages from the
specified arch. The "mirror" flag allows you to select which mirror to use.

We also remove checking for packages that overlap. Overlap checking was there to prevent downloading tarballs that are used by multiple modules. This is not really needed and is confusing when certian modules don't show up in the list.